Minor Class Refactoring
