+++ src/main/java/de/butzlabben/world/command/commands/WSCommands.java
@@ -1,6 +1,6 @@
 package de.butzlabben.world.command.commands;
 
-import de.butzlabben.world.WorldSystem;
+import de.butzlabben.WorldSystem;
 import de.butzlabben.world.autoupdater.AutoUpdater;
 import de.butzlabben.world.config.DependenceConfig;
 import de.butzlabben.world.config.MessageConfig;
