+++ src/main/java/de/butzlabben/world/command/commands/WorldAdministrateCommand.java
@@ -1,6 +1,6 @@
 package de.butzlabben.world.command.commands;
 
-import de.butzlabben.world.WorldSystem;
+import de.butzlabben.WorldSystem;
 import de.butzlabben.world.config.DependenceConfig;
 import de.butzlabben.world.config.MessageConfig;
 import de.butzlabben.world.config.PluginConfig;
