+++ src/main/java/de/butzlabben/world/util/VersionUtil.java
@@ -1,6 +1,5 @@
 package de.butzlabben.world.util;
 
-import de.butzlabben.world.WorldSystem;
 import org.bukkit.Bukkit;
 import org.bukkit.scheduler.BukkitTask;
